# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      metadata:
        additionalProperties:
          type: string
        type: object
      reason:
        type: string
    type: object
...
  dto.GetNotificationsResponse:
    properties:
      notifications:
        items:
          $ref: '#/definitions/dto.NotificationDTO'
        type: array
      total:
        example: 10
        type: integer
    type: object
...
  dto.NotificationDTO:
    properties:
      content:
        example: Math Quiz
        type: string
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_read:
        example: false
        type: boolean
      related_entity_id:
        description: 'For group_invite/group_kicked: group_id; for quiz_*: quiz instance
          id'
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      requires_action:
        example: false
        type: boolean
      title:
        example: New Quiz Available
        type: string
      type:
        enum:
        - group_invite
        - group_kicked
        - quiz_created
        - quiz_results
        - grade_changed
        - deadline_reminder
        example: quiz_created
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  dto.NotificationIDsRequest:
    properties:
      ids:
        example:
        - 550e8400-e29b-41d4-a716-446655440000
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
host: localhost:8080
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: Kollocol API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Get list of notifications for current user
      parameters:
      - default: 30
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetNotificationsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - notifications
  /notifications/delete:
    delete:
      consumes:
      - application/json
      description: Delete multiple notifications
      parameters:
      - description: Notification IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NotificationIDsRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete notifications
      tags:
      - notifications
  /notifications/read:
    put:
      consumes:
      - application/json
      description: Mark multiple notifications as read
      parameters:
      - description: Notification IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.NotificationIDsRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notifications as read
      tags:
      - notifications
  /notifications/read/all:
    put:
      description: Mark all of the current user's notifications as read (notifications
        requiring action are skipped)
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"